requirements.txt
sqlalchemy
python-dotenv
psycopg2-binary
Faker